Add support for Django 3.1
`widget_type` was introduced in Django 3.1, with a slightly different
syntax. We switch to this new syntax. The template tag can be deprecated
once Django 3.0 is no longer supported.
