+++ .travis.yml
@@ -22,6 +22,12 @@ matrix:
       env: TOXENV=py37-django30
     - python: 3.8
       env: TOXENV=py38-django30
+    - python: 3.6
+      env: TOXENV=py36-django31
+    - python: 3.7
+      env: TOXENV=py37-django31
+    - python: 3.8
+      env: TOXENV=py38-django31
 
     - python: 3.7
       env: TOXENV=lint
