+++ src/oscar/templates/oscar/partials/form_field.html
@@ -14,7 +14,7 @@
         <div class="form-group {% if field.errors %}has-error{% endif %}">
 
             {% block label %}
-                {% if not nolabel and field.widget_type != 'CheckboxInput' %}
+                {% if not nolabel and field.widget_type != 'checkbox' %}
                     <label for="{{ field.auto_id }}" class="{% if style|default:"stacked" != 'stacked' %}col-sm-4{% endif%} control-label{% if field.field.required %} required{% endif %}">
                         {{ field.label|safe }}
                     </label>
@@ -22,9 +22,9 @@
             {% endblock %}
 
             {% block controls %}
-                <div class="{% if style|default:"stacked" != 'stacked' %}col-sm-7{% endif %}{% if field.widget_type == 'CheckboxInput' %} checkbox{% endif %}">
+                <div class="{% if style|default:"stacked" != 'stacked' %}col-sm-7{% endif %}{% if field.widget_type == 'checkbox' %} checkbox{% endif %}">
                     {% block widget %}
-                        {% if field.widget_type == 'CheckboxInput' %}
+                        {% if field.widget_type == 'checkbox' %}
                             <label for="{{ field.auto_id }}" {% if field.field.required %}class="required"{% endif %}>
                                 {% render_field field %}
                                 {{ field.label|safe }}{% if field.field.required %} <span>*</span>{% endif %}
