+++ src/oscar/templatetags/form_tags.py
@@ -1,4 +1,8 @@
+import django
 from django import template
+from django.template.base import TextNode
+
+from oscar.core.compat import FormFieldNode
 
 register = template.Library()
 
@@ -9,21 +13,12 @@ def annotate_form_field(parser, token):
     Set an attribute on a form field with the widget type
 
     This means templates can use the widget type to render things differently
-    if they want to.  Django doesn't make this available by default.
+    if they want to. Until 3.1, Django did not make this available by default.
     """
     args = token.split_contents()
     if len(args) < 2:
         raise template.TemplateSyntaxError(
             "annotate_form_field tag requires a form field to be passed")
-    return FormFieldNode(args[1])
-
-
-class FormFieldNode(template.Node):
-    def __init__(self, field_str):
-        self.field = template.Variable(field_str)
-
-    def render(self, context):
-        field = self.field.resolve(context)
-        if hasattr(field, 'field'):
-            field.widget_type = field.field.widget.__class__.__name__
-        return ''
+    if django.VERSION < (3, 1):
+        return FormFieldNode(args[1])
+    return TextNode('')
