+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 envlist =
-    py{36,37,38}-django{22,30}
+    py{36,37,38}-django{22,30,31}
     lint
     sandbox
     docs
@@ -13,7 +13,7 @@ pip_pre = true
 deps =
     django22: django>=2.2,<2.3
     django30: django>=3.0,<3.1
-
+    django31: django>=3.1,<3.2
 
 [testenv:lint]
 basepython = python3.7
